+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 ## [Unreleased]
+## Fixed
+- Start time issue for Virtual parent Items, by @HardNorth
 
 ## [5.3.0]
 ### Added